% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Compute a genotype matrix and impute missing genotypes}
\usage{
impute(Data, impute = T, rmMono = T, rmMissing = T, method = NULL,
  genotyping = NULL, alpha = 0.05, verbose = T)
}
\arguments{
\item{Data}{A RUBIAS format data table. Missing values are NA. Alleles are coded as integers: A=1, C=2, G=3, and T=4 for SNPs.}

\item{impute}{Should missing genotypes be imputed. Default TRUE.}

\item{rmMono}{Should monomorphic SNPs be removed. Default TRUE.}

\item{rmMissing}{Shold SNPs with more than alpha percentage missing values be removed. Default TRUE.}

\item{method}{Imputation method used. If "mode" impute missing values with marker mode (row mode). If "random" impute missing values by selection of genotype based on the genotype frequencies across all populations. Default set as "mode".}

\item{genotyping}{The genotyping method used. If set as "Axiom", the number of reference alleles (C/G) is returned. If set as "Fluidigm", the original allele coding is preserved and four "genotypes" are returned (alleles are divided into independent observations at each locus).}

\item{alpha}{A user defined value: Remove genotypes with more than alpha percentage values missing. Default set as 5 percentage.}

\item{verbose}{Tell when it is done and everything went smoothly. Default TRUE.}
}
\value{
Returns a (nmb of SNPs)x(nmb of individuals) numeric valued data matrix. One should impute baseline and mixture data separately.
}
\description{
Compute a genotype matrix. One can also impute missing genotypes once either with marker mode (row mode) or random selection of gentype based on the genotype frequencies across the entire baseline.
}
\examples{
MixtureY = impute(MixtureData) # Simple marker mode imputation
ReferenceY = impute(ReferenceData)

# Choose only same set of markers:
setdiff(rownames(MixtureY),rownames(ReferenceY))
Markers = intersect(rownames(MixtureY),rownames(ReferenceY))

MixtureY = MixtureY[Markers,]
ReferenceY = ReferenceY[Markers,]

SampleSizeBaselinePop = ncol(ReferenceY)
Y = cbind(ReferenceY,MixtureY)
}
\keyword{Impute}
\keyword{genotypes}
\keyword{missing}
\keyword{monomorphic}
\keyword{values}
