% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSOSolPath.R
\name{LASSOSolPath}
\alias{LASSOSolPath}
\title{Compute LASSO solution path}
\usage{
LASSOSolPath(Y, lambda, intercept = T, SampleSizeBaselinePop,
  UseWeights = F, Baseline = T, verbose = T)
}
\arguments{
\item{Y}{A (nmb of SNPs)x(nmb of individuals) data matrix with numeric entries 0,1 and 2. No missing values allowed. First "SampleSizeBaselinePop" columns are SNPs of baseline individuals, rest are mixture individuals.}

\item{lambda}{A user defined vector of non-zero tuning parameter values in decreasing order.}

\item{intercept}{Should the intercept be included in the logit-model. Default TRUE.}

\item{SampleSizeBaselinePop}{How many baseline individuals are in data. A numeric value.}

\item{UseWeights}{Genotype weights (observation weights of the multinomial logit model). If TRUE, the procedure computes weights so that more weight is given to rare genotypes (computed in baseline and mixture group separately). When FALSE (default) weight is 1 for each genotype.}

\item{Baseline}{Should it also be checked how baseline individuals depend on mixture individual. Default TRUE.}

\item{verbose}{Should process be printed. The default value is TRUE.}
}
\value{
Returns an array of dimension [ncol(Y),ncol(Y),length(lambda)] and saves the array in the working directory. ith matrix (a non-symmetric adjacency matrix) in the array corresponds to an adjacency matrix associated with the tuning parameter value lambda_i.
}
\description{
Computes the LASSO solution path of the length length(lambda) using multinomial logit-model which is augmented with a LASSO-penalty function.
}
\examples{
MixtureY = impute(MixtureData) # Simple marker mode imputation
ReferenceY = impute(ReferenceData)

# Choose only same set of markers:
setdiff(rownames(MixtureY),rownames(ReferenceY))
Markers = intersect(rownames(MixtureY),rownames(ReferenceY))

MixtureY = MixtureY[Markers,]
ReferenceY = ReferenceY[Markers,]

SampleSizeBaselinePop = ncol(ReferenceY)
Y = cbind(ReferenceY,MixtureY)

lambda = seq(0.4,0.02,length.out=40)

MBapprox = LASSOSolPath(Y,lambda,intercept=T,SampleSizeBaselinePop,Baseline=T)
}
\keyword{Baseline}
\keyword{LASSO}
\keyword{Mixture}
\keyword{Network}
\keyword{Solution}
\keyword{glmnet}
\keyword{path}
