% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolPathInference.R
\name{SolPathInference}
\alias{SolPathInference}
\title{Solution path method}
\usage{
SolPathInference(MBapprox, ReferenceData, SampleNames,
  SampleSizeBaselinePop, alpha = 0.05)
}
\arguments{
\item{MBapprox}{A nonsymmetric network computed with the function LASSOSolPath.}

\item{ReferenceData}{A RUBIAS format baseline data file. Missing values are NA and A=1, C=2, G=3, and T=4 for SNPs.}

\item{SampleNames}{A character vector of sample names, both mixture and baseline: baseline first, then mixture.}

\item{SampleSizeBaselinePop}{A numeric value: how many baseline individuals there are?}

\item{alpha}{A alpha percentage quantile: individuals whose probability of the origin differs from the baseline population proportions less than or equal to this are determined as outgroup members. Note that there are always individuals which are determined as outgroup memebers! Exercise caution and more detailed examination.}
}
\value{
ProbofOrigin = a probability of the origin matrix where individuals are at rows and populations are columns. The last column is the MSE between individuals probability of the origin and baseline proportions. This table is also saved in a folder "NetworkResults" as "ProbOfOriginSolPath.txt"

MixtureProp = Mixture proportions. This table is also saved in a folder "NetworkResults" as "MixPropSolPath.txt"

OutgroupMembers = A vector of the potential outgroup memebers

NmbOfNeighbors = An array of dimension [nmb of baseline populations, nmb of mixture individuals, length(lambda)]. Each row represents the number of neighbors the individual has with the corresponding baseline population individuals.
}
\description{
Compute probability of the origin and mixture proportions using the whole LASSO solution path. Weighted averages are counted.
}
\examples{
MixtureY = impute(MixtureData) # Simple marker mode imputation
ReferenceY = impute(ReferenceData)

# Choose only same set of markers:
setdiff(rownames(MixtureY),rownames(ReferenceY))
Markers = intersect(rownames(MixtureY),rownames(ReferenceY))

MixtureY = MixtureY[Markers,]
ReferenceY = ReferenceY[Markers,]

SampleSizeBaselinePop = ncol(ReferenceY)
Y = cbind(ReferenceY,MixtureY)

lambda = seq(0.4,0.02,length.out=40)

MBapprox = LASSOSolPath(Y,lambda,intercept=T,SampleSizeBaselinePop,Baseline=T)

SampleNames=colnames(Y)

NetworkResultsSolpath = SolPathInference(MBapprox,ReferenceData,SampleNames=SampleNames,SampleSizeBaselinePop,alpha=0.05)
}
\keyword{Baseline}
\keyword{LASSO}
\keyword{Mixture}
\keyword{Network}
\keyword{Solution}
\keyword{path}
