% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTAInference.R
\name{WTAInference}
\alias{WTAInference}
\title{Winner takes it all method}
\usage{
WTAInference(MBapprox, ReferenceData, SampleNames, SampleSizeBaselinePop,
  NmbOfParents = 1)
}
\arguments{
\item{MBapprox}{A nonsymmetric network computed with the function LASSOSolPath.}

\item{ReferenceData}{A RUBIAS format baseline data file. Missing values are NA. Alleles are coded as integers: A=1, C=2, G=3, and T=4 for SNPs.}

\item{SampleNames}{A character vector of sample names, both mixture and baseline: baseline first, then mixture.}

\item{SampleSizeBaselinePop}{A numeric value: how many baseline individuals there are?}

\item{NmbOfParents}{How many neighbors should be found before we assing individual(s) to corresponding baseline populations? Default is 1.}
}
\value{
ProbofOrigin = A probability of the origin matrix where individuals are at rows and populations are columns. This table is also saved in a folder "NetworkResults" as "ProbOfOriginWTA.txt"

MixtureProp = Mixture proportions. This table is also saved in a folder "NetworkResults" as "MixPropWTA.txt".

OutgroupSample = IDs of mixture individuals which were not assigned to any baseline population. 

FirstLambda = The index of the penalty parameter lambda when NmbOfParents neighbors were found. This is NA if no neighbors is found.
}
\description{
By default, the condition is "associate the mixture individual to baseline population where the first neighbor is found"
}
\examples{
MixtureY = impute(MixtureData) # Simple marker mode imputation
ReferenceY = impute(ReferenceData)

# Choose only same set of markers:
setdiff(rownames(MixtureY),rownames(ReferenceY))
Markers = intersect(rownames(MixtureY),rownames(ReferenceY))

MixtureY = MixtureY[Markers,]
ReferenceY = ReferenceY[Markers,]

SampleSizeBaselinePop = ncol(ReferenceY)
Y = cbind(ReferenceY,MixtureY)

lambda = seq(0.4,0.02,length.out=40)

MBapprox = LASSOSolPath(Y,lambda,intercept=T,SampleSizeBaselinePop,Baseline=T)

SampleNames=colnames(Y)

NetworkResultsWTA = WTAInference(MBapprox,ReferenceData,SampleNames=SampleNames,SampleSizeBaselinePop,NmbOfParents=1)
}
\keyword{Baseline}
\keyword{LASSO}
\keyword{Network}
